onnxruntime: Fix Eigen SHA256 checksum mismatch during build
Issue: LIN1024-10399

When the machine-learning feature is enabled, the build fails due to a
SHA256 mismatch when fetching the Eigen source archive. This is likely
caused by upstream re-archiving of the same commit. Update the expected
checksum in onnxruntime to resolve the issue.

(LOCAL REV: NOT UPSTREAM) - Will send to meta-imx

Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>
Signed-off-by: Liyin Zhang <liyin.zhang.cn@windriver.com>